DR2-1383 Use new dynamo formatter classes
Because the validation we do against the dynamo classes depends on the
type of row, we've split out the formatters into different classes.

The problem here is that the children can be either folders or assets.
I've added a couple of custom case classes and a custom formatter which
will return the necessary fields from the asset or folder.

The validation for the asset will still happen, even though we aren't
using some of the validated fields.

The rest of the changes are updating the tests to add fields to the test
responses that are now mandatory