DR2-1434 Improve logging
This is similar to the other logging PRs. I haven't added the file
reference to the log context object as this lambda is for folders which
generally won't have it set. It is set on the asset but there may be
many assets to a folder so it's best to just leave it.
I've included the batch ref which can be used to track the ingest